% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\title{Get age from dob, accounting for leap years}
\usage{
age(dob, age_day = today(), units = "years", floor = TRUE)
}
\arguments{
\item{floor}{}
}
\description{
Get age from dob, accounting for leap years
}
