% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{calc_chisq}
\alias{calc_chisq}
\title{Calculate chisq and out-put a dataframe}
\usage{
calc_chisq(dat)
}
\arguments{
\item{dat}{results of a tablulation}
}
\value{
chisq tag
}
\description{
Calculate chisq and out-put a dataframe
}
