% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/utilities.R
\name{drop_extension}
\alias{drop_extension}
\title{Title}
\usage{
drop_extension(file_name)

drop_extension(file_name)
}
\arguments{
\item{file_name}{}

\item{file_name}{}
}
\description{
Title

Remove 3 letter extension from file name string
}

