% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/utilities.R
\name{folder_2stata}
\alias{folder_2stata}
\alias{folder_2stata}
\title{batch convert a folder of dataframes into individual stata files with
the same names}
\usage{
folder_2stata(dir_raw, dir_stata, format = "csv", col_widths = NULL)

folder_2stata(dir_raw, dir_stata, format = "csv", col_widths = NULL)
}
\arguments{
\item{dir_raw}{path to directory with files to be converted}

\item{dir_stata}{path to stata files directory}

\item{format}{valid formats are csv, xls, dta, tsv, sav}

\item{col_widths}{only used for reading fixed width files}

\item{dir_raw}{path to directory with files to be converted}

\item{dir_stata}{path to stata files directory}

\item{format}{valid formats are csv, xls, dta, tsv, sav}

\item{col_widths}{only used for reading fixed width files}
}
\description{
batch convert a folder of dataframes into individual stata files with
the same names

batch convert a folder of dataframes into individual stata files with
the same names
}
