% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/utilities.R
\name{folder_bind}
\alias{folder_bind}
\alias{folder_bind}
\title{Bind a folder of separate data files with the same headings into one object
takes excel, stata, spss, fixed width files, and standard table formats}
\usage{
folder_bind(dir_raw, format = "csv", col_widths = NULL)

folder_bind(dir_raw, format = "csv", col_widths = NULL)
}
\arguments{
\item{format}{valid formats are csv, xls, dta, tsv, sav}

\item{col_widths}{only used for reading fixed width files}

\item{path}{the path to the folder where the data files are kept}

\item{path}{the path to the folder where the data files are kept}

\item{format}{valid formats are csv, xls, dta, tsv, sav}

\item{col_widths}{only used for reading fixed width files}
}
\value{
bound data from

bound data from
}
\description{
Bind a folder of separate data files with the same headings into one object
takes excel, stata, spss, fixed width files, and standard table formats

Bind a folder of separate data files with the same headings into one object
takes excel, stata, spss, fixed width files, and standard table formats
}
