% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_contacts.R
\name{generate_posts}
\alias{generate_posts}
\title{Generate a list of json payloads for put requests}
\usage{
generate_posts(df_new_ids)
}
\arguments{
\item{df_new_ids}{}
}
\description{
Generate a list of json payloads for put requests
}
