% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mailing_list.R
\name{get_contacts}
\alias{get_contacts}
\title{Pull information from a contact list. Currently
just for OBIP participants}
\usage{
get_contacts(list_id)
}
\arguments{
\item{list_id}{}
}
\description{
Pull information from a contact list. Currently
just for OBIP participants
}
