% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mailing_list.R
\name{get_mailing_lists}
\alias{get_mailing_lists}
\title{xcollect all mailing lists associated with teh}
\usage{
get_mailing_lists(search = NULL)
}
\arguments{
\item{search}{}
}
\description{
xcollect all mailing lists associated with teh
}
