% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mailing_list.R
\name{get_page_list}
\alias{get_page_list}
\title{List urls for each page of the contact list}
\usage{
get_page_list(list_id)
}
\arguments{
\item{list_id}{}
}
\description{
List urls for each page of the contact list
}
