% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/utilities.R
\name{get_reader}
\alias{get_reader}
\alias{get_reader}
\title{Get a reader function for a file format}
\usage{
get_reader(format)

get_reader(format)
}
\arguments{
\item{format}{}

\item{format}{}
}
\value{
reader_function

reader_function
}
\description{
Get a reader function for a file format

Get a reader function for a file format
}
