% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_utils.R
\name{get_survey}
\alias{get_survey}
\title{Get an export of responses from a survey}
\usage{
get_survey(id, folder = "Z:/R/temp", fname = "qxre.zip", format = "spss",
  labs = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Get an export of responses from a survey
}
