% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{gg_barbell}
\alias{gg_barbell}
\title{make a brett-lamb style barbell chart, with either the whole
bar from 0 indicated, or just the difference.}
\usage{
gg_barbell(dat, x, xend, y, m, full_bar = FALSE, pct = FALSE)
}
\arguments{
\item{pct}{}
}
\description{
make a brett-lamb style barbell chart, with either the whole
bar from 0 indicated, or just the difference.
}
