% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{gg_change}
\alias{gg_change}
\title{plot a variable as change across facets}
\usage{
gg_change(dat, x, y, m, f, lab, scales)
}
\arguments{
\item{scales}{}
}
\description{
plot a variable as change across facets
}
