% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{\%!find\%}
\alias{\%!find\%}
\title{Find observations in a vector that are not NA and don't match a value}
\usage{
x \%!find\% y
}
\arguments{
\item{y}{}
}
\description{
Find observations in a vector that are not NA and don't match a value
}
