% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Inverse %in% operator}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{the thing}

\item{y}{what it is / isn't in}
}
\value{
logical vector, true when x is not in y
}
\description{
Inverse %in% operator
}
