% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kablify.R
\name{kablify}
\alias{kablify}
\title{Build a kable from a table, change proportions to
percentages}
\usage{
kablify(tab)
}
\arguments{
\item{tab}{}
}
\description{
Build a kable from a table, change proportions to
percentages
}
