% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/utilities.R
\name{list_used_files}
\alias{list_used_files}
\title{returns list of files of a given format}
\usage{
list_used_files(dir_raw, format)

list_used_files(dir_raw, format)
}
\arguments{
\item{format}{format of files to filter for}

\item{dir_path}{path to the directory of files to filter}

\item{dir_path}{string: path to the directory of files to filter}

\item{format}{string: format of files to filter for}
}
\value{
character vector of files
}
\description{
returns list of files of a given format

Get list of files to be processed
}

