% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_multitable.R
\name{multi_table}
\alias{multi_table}
\title{Write a named list of tables to a csv file with headers}
\usage{
multi_table(dat, file_path)
}
\arguments{
\item{file_path}{}
}
\description{
Write a named list of tables to a csv file with headers
}
