% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_contacts.R
\name{post_contacts}
\alias{post_contacts}
\title{Post a dataframe of new contacts to the contact list}
\usage{
post_contacts(new_contacts, list_id)
}
\arguments{
\item{posts}{}
}
\description{
Post a dataframe of new contacts to the contact list
}
