% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/utilities.R
\name{read_raw_excel}
\alias{read_raw_excel}
\alias{read_raw_excel}
\title{Title}
\usage{
read_raw_excel(file_name)

read_raw_excel(file_name)
}
\arguments{
\item{file_name}{}

\item{file_name}{}
}
\description{
Title

Automatically detect excel header
}
