% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_survey.R
\name{read_survey}
\alias{read_survey}
\title{Get response data using a qualtrics survey id}
\usage{
read_survey(survey_id, labs = FALSE, ...)
}
\arguments{
\item{survey_id}{}
}
\description{
Get response data using a qualtrics survey id
}
