% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{rts_eia_inv}
\alias{rts_eia_inv}
\title{Read EIA involvement raw files into one data frame}
\usage{
rts_eia_inv(dir_eia, dir_stata)
}
\arguments{
\item{dir_eia}{directory with EIA files}
}
\value{
df_eia_inv concatenated data frame
}
\description{
Read EIA involvement raw files into one data frame
}

