% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_replace_when.R
\name{str_replace_when}
\alias{str_replace_when}
\title{Replace substrings in a vector sequentially, with arguments structured
like case_when()}
\usage{
str_replace_when(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Replace substrings in a vector sequentially, with arguments structured
like case_when()
}
\examples{


}
