% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Workflow / print / NSE friendly implementation of table()}
\usage{
tab(dat, x, y, n = FALSE, col = FALSE, row = FALSE, cell = FALSE,
  row_p = FALSE, print = TRUE, total_row = FALSE, chisq = TRUE)
}
\arguments{
\item{cell}{}
}
\description{
Workflow / print / NSE friendly implementation of table()
}
