% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_list.R
\name{tabset}
\alias{tabset}
\title{Make a set of tab functions to map onto any data with appropriate colnames}
\usage{
tabset(y, ..., by = none)
}
\arguments{
\item{...}{}
}
\description{
Make a set of tab functions to map onto any data with appropriate colnames
}
