% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mailing_list.R
\name{tabulate_request}
\alias{tabulate_request}
\title{Tabulate the results of a get request}
\usage{
tabulate_request(url)
}
\arguments{
\item{url}{}
}
\description{
Tabulate the results of a get request
}
