% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_list.R
\name{write_tab_list}
\alias{write_tab_list}
\title{Print a list of tables as a writable vector, separated by third level headers}
\usage{
write_tab_list(tab_list)
}
\arguments{
\item{tab_list}{}
}
\description{
Print a list of tables as a writable vector, separated by third level headers
}
