% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_options.R
\name{register_options}
\alias{register_options}
\title{Wrapper for qualtRics::registerOptions which includes
our root url and the location of the encrypted API key}
\usage{
register_options()
}
\description{
Wrapper for qualtRics::registerOptions which includes
our root url and the location of the encrypted API key
}
